Call composer command to copy .env file
If you run the installer to create a new project it says the
Application key was set successfully even though it wasn't.

This calls the composer script to copy .env.example to .env, so the
next script that calls artisan key:generate will work.

The reason for the current behavior is that php artisan key:generate
doesn't display an error if it fails to find a .env file.