Update NewCommand.php
use the `extension_loaded` method to check if the 'zip' extension is loaded rather than checking for a class in the extension.

while probably an edge case, a their could be a user defined `\ZipArchive` class that would cause this conditional to pass, even when the extension is not loaded.